Windows: Prevent git from changing line endings on test files
The sample3.ref test file is plain ascii, which means that Git on
Windows will automatically convert the line endings from \n to \r\n.
This of course causes a test to fail.

Added a .gitattributes file that requires Git to treat *.ref files as
binary, so it won't convert line endings.